+++ src/main/java/frc/robot/RobotContainer.java
@@ -11,9 +11,6 @@ import edu.wpi.first.wpilibj.GenericHID;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.commands.ArcadeDrive;
-import frc.robot.commands.PositionForHub;
-import frc.robot.commands.RotateToHub;
-import frc.robot.commands.VisionSniperMode;
 import frc.robot.commands.auto.FourCargoAuto;
 import frc.robot.commands.operatorcommands.TeleClimbDown;
 import frc.robot.commands.operatorcommands.TeleClimbTilt;
@@ -26,6 +23,9 @@ import frc.robot.commands.operatorcommands.TeleIndexer;
 import frc.robot.commands.operatorcommands.TeleIntake;
 import frc.robot.commands.operatorcommands.TeleIntakeToggle;
 import frc.robot.commands.operatorcommands.TeleUptake;
+import frc.robot.commands.visioncommands.PositionForHub;
+import frc.robot.commands.visioncommands.RotateToHub;
+import frc.robot.commands.visioncommands.VisionSniperMode;
 import frc.robot.oi.MixedXboxJoystickControls;
 import frc.robot.oi.UserControls;
 import frc.robot.subsystems.climber.Climber;
