+++ src/main/java/frc/robot/commands/auto/FourCargoAuto.java
@@ -4,13 +4,13 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.commands.DriveStraight;
-import frc.robot.commands.PositionForHub;
 import frc.robot.commands.RunFlyWheel;
 import frc.robot.commands.SetIntakeExtended;
 import frc.robot.commands.Turn;
 import frc.robot.commands.operatorcommands.TeleFeed;
 import frc.robot.commands.operatorcommands.TeleIndexer;
 import frc.robot.commands.operatorcommands.TeleIntake;
+import frc.robot.commands.visioncommands.PositionForHub;
 import frc.robot.subsystems.colorSensor.ColorSensor;
 import frc.robot.subsystems.drivetrain.DriveBase;
 import frc.robot.subsystems.flywheel.VarFlyWheel;
