+++ src/main/java/frc/robot/commands/visioncommands/VisionSniperMode.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.visioncommands;
 
 import java.util.function.Supplier;
